- Separate training logic from "posterior" logic, with two object: `Posterior` and `Trainer`.
- remove data_loaders.py and include data logic in Trainer.
