+++ scvi/models/__init__.py
@@ -2,7 +2,9 @@ from .svaec import SVAEC
 from .vae import VAE
 from .vaec import VAEC
 from .vae_fish import VAEF
+from.classifier import Classifier
 __all__ = ['VAEC',
            'SVAEC',
            'VAE',
-           'VAEF']
+           'VAEF',
+           'Classifier']
